Fixed circular input
