+++ trackers/utils/bbox_conversions.py
@@ -1,0 +1,34 @@
+import numpy as np
+
+
+def convert_bbox_to_state_rep(bbox: np.ndarray) -> np.ndarray:
+    """
+    Takes a bounding box in the form [x1,y1,x2,y2] and returns z in the form
+      [x,y,s,r] where x,y is the centre of the box and s is the scale/area and r is
+      the aspect ratio
+    Args:
+        bbox: bounding box in the form [x1,y1,x2,y2]
+    Returns:
+        np.ndarray: bounding box in the form [x,y,s,r]
+    """
+    w = bbox[2] - bbox[0]
+    h = bbox[3] - bbox[1]
+    x = bbox[0] + w / 2.0
+    y = bbox[1] + h / 2.0
+    s = w * h  # scale is just area
+    r = w / float(h + 1e-6)
+    return np.array([x, y, s, r])
+
+
+def convert_state_rep_to_bbox(x):
+    """
+    Takes a bounding box in the centre form [x,y,s,r] and returns it in the form
+      [x1,y1,x2,y2] where x1,y1 is the top left and x2,y2 is the bottom right
+    Args:
+        bbox: bounding box in the form [x,y,s,r]
+    Returns:
+        np.ndarray: bounding box in the form [x1,y1,x2,y2]
+    """
+    w = np.sqrt(x[2] * x[3])
+    h = x[2] / w
+    return np.array([x[0] - w / 2.0, x[1] - h / 2.0, x[0] + w / 2.0, x[1] + h / 2.0])
